requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
groq==0.4.2
python-dotenv==1.0.0
plotly==5.17.0
altair==5.1.2
fastapi==0.110.2
uvicorn==0.29.0
langgraph==0.0.40
psycopg2-binary==2.9.9 # For optional PostgreSQL support 